Move modules outside generic protocols
