+++ codegen-server/src/main/kotlin/software/amazon/smithy/rust/codegen/server/smithy/generators/protocol/ServerProtocol.kt
@@ -99,10 +99,10 @@ class ServerAwsJsonProtocol(
     override fun markerStruct(): RuntimeType {
         return when (version) {
             is AwsJsonVersion.Json10 -> {
-                ServerRuntimeType.Protocol("AwsJson10", "aws_json::aws_json_10", runtimeConfig)
+                ServerRuntimeType.Protocol("AwsJson10", "aws_json_10", runtimeConfig)
             }
             is AwsJsonVersion.Json11 -> {
-                ServerRuntimeType.Protocol("AwsJson11", "aws_json::aws_json_11", runtimeConfig)
+                ServerRuntimeType.Protocol("AwsJson11", "aws_json_11", runtimeConfig)
             }
         }
     }
@@ -211,7 +211,7 @@ class ServerRestJsonProtocol(
         fun fromCoreProtocol(restJson: RestJson): ServerRestJsonProtocol = ServerRestJsonProtocol(restJson.coreCodegenContext)
     }
 
-    override fun markerStruct() = ServerRuntimeType.Protocol("AwsRestJson1", "rest::rest_json_1", runtimeConfig)
+    override fun markerStruct() = ServerRuntimeType.Protocol("AwsRestJson1", "rest_json_1", runtimeConfig)
 
     override fun routerType() = restRouterType(runtimeConfig)
 
@@ -240,7 +240,7 @@ class ServerRestXmlProtocol(
         }
     }
 
-    override fun markerStruct() = ServerRuntimeType.Protocol("AwsRestXml", "rest::rest_xml", runtimeConfig)
+    override fun markerStruct() = ServerRuntimeType.Protocol("AwsRestXml", "rest_xml", runtimeConfig)
 
     override fun routerType() = restRouterType(runtimeConfig)
