+++ rust-runtime/aws-smithy-http-server/src/proto/aws_json/mod.rs
@@ -3,6 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod aws_json_10;
-pub mod aws_json_11;
 pub mod router;
