+++ rust-runtime/aws-smithy-http-server/src/proto/mod.rs
@@ -4,4 +4,8 @@
  */
 
 pub mod aws_json;
+pub mod aws_json_10;
+pub mod aws_json_11;
 pub mod rest;
+pub mod rest_json_1;
+pub mod rest_xml;
