+++ rust-runtime/aws-smithy-http-server/src/proto/rest/mod.rs
@@ -3,6 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-pub mod rest_json_1;
-pub mod rest_xml;
 pub mod router;
