+++ rust-runtime/aws-smithy-http-server/src/proto/rest_xml/mod.rs
