+++ rust-runtime/aws-smithy-http-server/src/routing/mod.rs
@@ -16,8 +16,8 @@ use self::request_spec::RequestSpec;
 use crate::{
     body::{boxed, Body, BoxBody, HttpBody},
     proto::{
-        aws_json::{aws_json_10::AwsJson10, aws_json_11::AwsJson11, router::AwsJsonRouter},
-        rest::{rest_json_1::AwsRestJson1, rest_xml::AwsRestXml, router::RestRouter},
+        aws_json::router::AwsJsonRouter, aws_json_10::AwsJson10, aws_json_11::AwsJson11, rest::router::RestRouter,
+        rest_json_1::AwsRestJson1, rest_xml::AwsRestXml,
     },
 };
 use crate::{error::BoxError, routers::RoutingService};
